server/requirements.txt
fastapi==0.112.0
uvicorn==0.30.1
pydantic==2.8.2
python-multipart==0.0.9

SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==1.0.1

alembic==1.13.2

python-dotenv==1.0.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4